% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/accessors_public.R
\name{GeneRankFit-accessors}
\alias{GeneRankFit-accessors}
\alias{params}
\alias{params,GeneRankFit-method}
\alias{oof}
\alias{oof,GeneRankFit-method}
\alias{imp}
\alias{imp,GeneRankFit-method}
\alias{imp<-}
\alias{imp<-,GeneRankFit-method}
\alias{calibration}
\alias{calibration,GeneRankFit-method}
\alias{calibration<-}
\alias{calibration<-,GeneRankFit-method}
\title{Accessors for GeneRankFit}
\usage{
params(x)

\S4method{params}{GeneRankFit}(x)

oof(x)

\S4method{oof}{GeneRankFit}(x)

imp(x)

\S4method{imp}{GeneRankFit}(x)

imp(x) <- value

\S4method{imp}{GeneRankFit}(x) <- value

calibration(x)

\S4method{calibration}{GeneRankFit}(x)

calibration(x) <- value

\S4method{calibration}{GeneRankFit}(x) <- value
}
\arguments{
\item{x}{A \code{GeneRankFit} object.}

\item{value}{Replacement value.}
}
\value{
For getters:
\describe{
\item{\code{params(x)}}{A list of training parameters and metadata.}
\item{\code{oof(x)}}{A list of out-of-fold results including predictions and labels.}
\item{\code{imp(x)}}{A \code{data.frame} containing aggregated variable importance scores.}
\item{\code{calibration(x)}}{A list describing calibration settings.}
}

For setters:
\describe{
\item{\code{imp(x) <- value}}{Returns the updated \code{GeneRankFit} object with modified importance data.}
\item{\code{calibration(x) <- value}}{Returns the updated \code{GeneRankFit} object with modified calibration settings.}
}
}
\description{
Getter/setter accessors for \code{GeneRankFit} slots.
}
\examples{
fit <- methods::new("GeneRankFit")

# getters
params(fit)
oof(fit)
imp(fit)
calibration(fit)

# setters
imp(fit) <- data.frame(
  gene = c("GeneA", "GeneB"),
  importance = c(0.8, 0.3)
)
calibration(fit) <- list(
  method = "platt",
  fun = function(p) p
)

imp(fit)
calibration(fit)

}
