% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package-doc.R
\docType{package}
\name{RFGeneRank}
\alias{RFGeneRank-package}
\alias{RFGeneRank}
\title{RFGeneRank: CV-stable predictive ranking for transcriptomics}
\description{
Package-level documentation and imports.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/USER/RFGeneRank}
  \item Report bugs at \url{https://github.com/USER/RFGeneRank/issues}
}

}
\author{
\strong{Maintainer}: Abdulaziz Albeshri \email{a.z.a1410@hotmail.com} (\href{https://orcid.org/0000-0002-0718-2439}{ORCID})

Other contributors:
\itemize{
  \item Thamer Ahmad Bouback \email{Tbouback@kau.edu.sa} [contributor]
  \item Majid Al-Zahrani \email{maalzahrani4@kau.edu.sa} [contributor]
  \item Tasneem Alsahafi \email{Talsahafi0015@stu.kau.edu.sa} [contributor]
}

}
\keyword{internal}
